change size_t to uint64_t

+++ src/caches.c
@@ -9,7 +9,7 @@ ternaryopcache_t ternaryopcache;
 
 void init_unaryopcache(uint64_t log2size) {
     assert(log2size < 32);
-    uint32_t size = (1 << log2size);
+    uint32_t size = ((uint32_t) 1) << log2size;
     unaryopcache.size = size;
     unaryopcache.data = (unaryopcache_entry_t*) malloc(size * sizeof(unaryopcache_entry_t));
     if (unaryopcache.data == NULL) {
@@ -17,7 +17,7 @@ void init_unaryopcache(uint64_t log2size) {
         assert(0);
     }
     unaryopcache.mask = size - 1;
-    for (int i = 0; i < size; i++) {
+    for (uint32_t i = 0; i < size; i++) {
         unaryopcache.data[i].arg = 0;
         unaryopcache.data[i].result = (nodeindex_t) -1;
         unaryopcache.data[i].op = UNDEFINED_OP;
@@ -27,14 +27,14 @@ inline uint32_t unaryop_hash(nodeindex_t ix, uint8_t op) {
     return hash_2(op, ix);
 }
 void clear_unaryopcache() {
-    for (int i = 0; i < unaryopcache.size; i++) {
+    for (uint32_t i = 0; i < unaryopcache.size; i++) {
         unaryopcache.data[i].op = UNDEFINED_OP;
     }
 }
 
 void init_binaryopcache(uint64_t log2size) {
     assert(log2size < 32);
-    uint32_t size = (1 << log2size);
+    uint32_t size = ((uint32_t) 1) << log2size;
     binaryopcache.size = size;
     binaryopcache.data = (binaryopcache_entry_t*) malloc(size * sizeof(binaryopcache_entry_t));
     if (binaryopcache.data == NULL) {
@@ -42,7 +42,7 @@ void init_binaryopcache(uint64_t log2size) {
         assert(0);
     }
     binaryopcache.mask = size - 1;
-    for (int i = 0; i < size; i++) {
+    for (uint32_t i = 0; i < size; i++) {
         binaryopcache.data[i].arg1 = 0;
         binaryopcache.data[i].arg2 = 0;
         binaryopcache.data[i].result = (nodeindex_t) -1;
@@ -54,14 +54,14 @@ inline uint32_t binaryop_hash(nodeindex_t ix1, nodeindex_t ix2, uint8_t op) {
     return hash_3(op, ix1, ix2);
 }
 void clear_binaryopcache() {
-    for (int i = 0; i < binaryopcache.size; i++) {
+    for (uint32_t i = 0; i < binaryopcache.size; i++) {
         binaryopcache.data[i].op = UNDEFINED_OP;
     }
 }
 
 void init_ternaryopcache(uint64_t log2size) {
     assert(log2size < 32);
-    uint32_t size = (1 << log2size);
+    uint32_t size = ((uint32_t) 1) << log2size;
     ternaryopcache.size = size;
     ternaryopcache.data = (ternaryopcache_entry_t*) malloc(size * sizeof(ternaryopcache_entry_t));
     if (binaryopcache.data == NULL) {
@@ -69,7 +69,7 @@ void init_ternaryopcache(uint64_t log2size) {
         assert(0);
     }
     ternaryopcache.mask = size - 1;
-    for (int i = 0; i < size; i++) {
+    for (uint32_t i = 0; i < size; i++) {
         ternaryopcache.data[i].arg1 = 0;
         ternaryopcache.data[i].arg2 = 0;
         ternaryopcache.data[i].arg3 = 0;
@@ -84,7 +84,7 @@ inline uint32_t ternaryop_hash(nodeindex_t ix1, nodeindex_t ix2, nodeindex_t ix3
     return hash_3(op, ix1, ix2);
 }
 void clear_ternaryopcache() {
-    for (int i = 0; i < ternaryopcache.size; i++) {
+    for (uint32_t i = 0; i < ternaryopcache.size; i++) {
         ternaryopcache.data[i].op = UNDEFINED_OP;
     }
 }
